CF-289 organize forecast packages
