+++ ta4j-core/src/main/java/org/ta4j/core/indicators/forecast/adapters/package-info.java
@@ -1,0 +1,12 @@
+/*
+ * SPDX-License-Identifier: MIT
+ */
+/**
+ * Forecast conversion adapters.
+ * <p>
+ * Adapter indicators bridge forecast domains or API shapes, for example by
+ * converting cumulative log-return forecasts into price forecasts.
+ *
+ * @since 0.22.9
+ */
+package org.ta4j.core.indicators.forecast.adapters;
