+++ ta4j-core/src/main/java/org/ta4j/core/indicators/forecast/projection/ReturnForecastProjectionIndicator.java
@@ -1,7 +1,7 @@
 /*
  * SPDX-License-Identifier: MIT
  */
-package org.ta4j.core.indicators.forecast;
+package org.ta4j.core.indicators.forecast.projection;
 
 import org.ta4j.core.criteria.ReturnRepresentation;
