+++ ta4j-core/src/main/java/org/ta4j/core/indicators/forecast/state/package-info.java
@@ -1,0 +1,13 @@
+/*
+ * SPDX-License-Identifier: MIT
+ */
+/**
+ * Hidden forecast state contracts and data records.
+ * <p>
+ * State indicators estimate latent inputs, such as rolling return mean and
+ * volatility, that projection indicators can consume without coupling to a
+ * specific estimator implementation.
+ *
+ * @since 0.22.9
+ */
+package org.ta4j.core.indicators.forecast.state;
