+++ ta4j-core/src/test/java/org/ta4j/core/indicators/forecast/state/ReturnForecastStateTest.java
@@ -1,7 +1,7 @@
 /*
  * SPDX-License-Identifier: MIT
  */
-package org.ta4j.core.indicators.forecast;
+package org.ta4j.core.indicators.forecast.state;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertThrows;
